Renamed files to new naming standard
