+++ src/models/events/follow-event.ts
@@ -1,7 +1,7 @@
 import { FieldStore, FieldKeys } from '@utilities';
 
-import { StreamEvent } from './streamEvent';
-import { StreamEventType } from './streamEventType';
+import { StreamEvent } from './stream-event';
+import { StreamEventType } from './stream-event-type';
 
 export class FollowEvent extends StreamEvent {
     public alertSound: string;
