+++ src/models/streamElements/event-received-detail.ts
@@ -1,0 +1,10 @@
+import { EventData } from './event-data';
+
+/**
+ * A contract for the detail of an 'onEventReceived' event.
+ */
+export interface EventReceivedDetail {
+    event: EventData;
+
+    listener: string;
+}
