--- src/models/streamElements/eventReceivedDetail.ts
@@ -1,10 +1,0 @@
-import { EventData } from './eventData';
-
-/**
- * A typed contract of the detail of a 'onEventReceived' event.
- */
-export interface EventReceivedDetail {
-    event: EventData;
-
-    listener: string;
-}
