+++ src/utilities/index.ts
@@ -1,3 +1,3 @@
-export { AlertHelper } from './alertHelper';
-export { FieldKeys, FieldStore } from './fieldStore';
+export { AlertHelper } from './alert-helper';
+export { FieldKeys, FieldStore } from './field-store';
 export { Time } from './time';
